Add base class for FlatMap*GroupsIn*Exec
