+++ sql/core/src/main/scala/org/apache/spark/sql/execution/python/FlatMapGroupsInArrowExec.scala
@@ -17,15 +17,11 @@
 
 package org.apache.spark.sql.execution.python
 
-import org.apache.spark.api.python.{ChainedPythonFunctions, PythonEvalType}
-import org.apache.spark.rdd.RDD
+import org.apache.spark.api.python.PythonEvalType
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions._
-import org.apache.spark.sql.catalyst.plans.physical.{AllTuples, ClusteredDistribution, Distribution, Partitioning}
-import org.apache.spark.sql.execution.{SparkPlan, UnaryExecNode}
-import org.apache.spark.sql.execution.python.PandasGroupUtils._
+import org.apache.spark.sql.execution.SparkPlan
 import org.apache.spark.sql.types.{StructField, StructType}
-import org.apache.spark.sql.util.ArrowUtils
 
 
 /**
@@ -50,54 +46,18 @@ case class FlatMapGroupsInArrowExec(
     func: Expression,
     output: Seq[Attribute],
     child: SparkPlan)
-  extends SparkPlan with UnaryExecNode with PythonSQLMetrics {
+  extends FlatMapGroupsInPythonExec {
 
-  private val sessionLocalTimeZone = conf.sessionLocalTimeZone
-  private val pythonRunnerConf = ArrowUtils.getPythonRunnerConfMap(conf)
-  private val pandasFunction = func.asInstanceOf[PythonUDF].func
-  private val chainedFunc = Seq(ChainedPythonFunctions(Seq(pandasFunction)))
+  protected val pythonEvalType: Int = PythonEvalType.SQL_GROUPED_MAP_ARROW_UDF
 
-  override def producedAttributes: AttributeSet = AttributeSet(output)
+  override protected def groupedData(iter: Iterator[InternalRow], attrs: Seq[Attribute]):
+      Iterator[Iterator[InternalRow]] =
+    super.groupedData(iter, attrs)
+      // Here we wrap it via another row so that Python sides understand it as a DataFrame.
+      .map(_.map(InternalRow(_)))
 
-  override def outputPartitioning: Partitioning = child.outputPartitioning
-
-  override def requiredChildDistribution: Seq[Distribution] = {
-    if (groupingAttributes.isEmpty) {
-      AllTuples :: Nil
-    } else {
-      ClusteredDistribution(groupingAttributes) :: Nil
-    }
-  }
-
-  override def requiredChildOrdering: Seq[Seq[SortOrder]] =
-    Seq(groupingAttributes.map(SortOrder(_, Ascending)))
-
-  override protected def doExecute(): RDD[InternalRow] = {
-    val inputRDD = child.execute()
-
-    val (dedupAttributes, argOffsets) = resolveArgOffsets(child.output, groupingAttributes)
-
-    // Map grouped rows to ArrowPythonRunner results, Only execute if partition is not empty
-    inputRDD.mapPartitionsInternal { iter => if (iter.isEmpty) iter else {
-
-      val data = groupAndProject(iter, groupingAttributes, child.output, dedupAttributes)
-        .map { case (_, x) => x }
-        // Here we wrap it via another row so that Python sides understand it
-        // as a DataFrame.
-        .map(_.map(InternalRow(_)))
-
-      val runner = new ArrowPythonRunner(
-        chainedFunc,
-        PythonEvalType.SQL_GROUPED_MAP_ARROW_UDF,
-        Array(argOffsets),
-        StructType(StructField("struct", StructType.fromAttributes(dedupAttributes)) :: Nil),
-        sessionLocalTimeZone,
-        pythonRunnerConf,
-        pythonMetrics)
-
-      executePython(data, output, runner)
-    }}
-  }
+  override protected def groupedSchema(attrs: Seq[Attribute]): StructType =
+    StructType(StructField("struct", super.groupedSchema(attrs)) :: Nil)
 
   override protected def withNewChildInternal(newChild: SparkPlan): FlatMapGroupsInArrowExec =
     copy(child = newChild)
